Add MacOS Omnisharp custom run script
This is a custom version of the Omnisharp run script designed to work around specific problems I was running into on my Mac trying to get Unity intellisense in VSCode. The major problem was that Omnisharp refused to find my global mono install - despite it being in the path. Interestingly, Omnisharp is itself launched within mono on Mac. This script only ran the mono built into Omnisharp, which caused the major problem. This change results in the use of a global mono if that is 6.4.0 or later, which mirrors Omnisharp's behavior when it picks the proper version of MSBuild.